Add basic memory representation of an int array
Add a previously written function that writes a memory representation
of an arbitrary data type (void *) to the standard output.

Each line represents sixteen bytes, first displayed in hexadecimal
format, then as ASCII unsigned chars. Any byte that corresponds to a
non-printable symbol in ASCII gets displayed as a period.

As a test value, the function is passed an int array.